requirements.txt
flask
flask_sqlalchemy
sqlalchemy
authlib
pyjwt
pika
python-dotenv
pymysql
paho-mqtt
requests
pyjwt